requirements.txt
asyncio
websockets>=12.0
RPi.GPIO
spidev